improve avatar loading
When avatar parts are not available but an uuid is, the Avatar item
will try to load the list of parts by itself. It’s useful when displaying
penpals because old conversations do to not have userStyles
(from which are derived avatar parts) included and thus penpals aren’t
populated with them.